+++ examples/demo11.py
@@ -118,13 +118,13 @@ def main(argv):
  
     # Read in galaxy catalog
     # The COSMOSCatalog uses the same input file as we have been usign for RealGalaxyCatalogs
-    # along with a second file called real_galaxy_catalog_examples_fits.fits, which stores
+    # along with a second file called real_galaxy_catalog_23.5_examples_fits.fits, which stores
     # the information about the parameteric fits.  There is no need to specify the second file
     # name, since the name is derivable from the name of the main catalog.
     if True:
         # The catalog we distribute with the GalSim code only has 100 galaxies.
         # The galaxies will typically be reused several times here.
-        cat_file_name = 'real_galaxy_catalog_example.fits'
+        cat_file_name = 'real_galaxy_catalog_23.5_example.fits'
         dir = 'data'
         cosmos_cat = galsim.COSMOSCatalog(cat_file_name, dir=dir)
     else: